refactor: extract CallingContext to top-level public class
- Create CallingContext.java as a separate top-level class
- Update CLIProcessor to use the extracted CallingContext
- Add getOutputStream() method to CLIProcessor for CallingContext access
- Make OPTIONS and handleNoColor/handleQuiet package-private
- Update all imports from CLIProcessor.CallingContext to CallingContext
- Fixes part of #252
