+++ cli-processor/src/main/java/gov/nist/secauto/metaschema/cli/processor/command/ICommandExecutor.java
@@ -5,7 +5,7 @@
 
 package gov.nist.secauto.metaschema.cli.processor.command;
 
-import gov.nist.secauto.metaschema.cli.processor.CLIProcessor.CallingContext;
+import gov.nist.secauto.metaschema.cli.processor.CallingContext;
 
 import org.apache.commons.cli.CommandLine;
